+++ artworks/constants.py
@@ -1,6 +1,6 @@
 """Constants."""
 
-BATCH_SIZE = 3
+BATCH_SIZE = 10
 MONGODB_URL = "mongodb+srv://artist:music@bmat.jl51a.mongodb.net/artworks?retryWrites=true&w=majority"
 DATASET = 'artworks_db'
 COLLECTION = 'artworks'