+++ docs/changelog.rst
@@ -4,7 +4,8 @@ Changelog
 [Unreleased]
 ------------
 * Read from CSV and insert into MongoDB script
-* API with GET route to get right owners metadata
+* API with GET routes: to get right owners metadata and all artworks
+* Documentation
 
 [0.0.1] - 2021-03-02
 --------------------